refactor(core): address code review findings
- Fix #1: Wrap decodeURIComponent in try/catch to handle malformed URIs
- Fix #2: Extract internal API routes to separate internal-api.ts module
- Fix #3: Add JSDoc documentation clarifying updateSeeds behavior
- Fix #4: Version constant defined in internal-api.ts (PACKAGE_VERSION)
- Fix #5: Add test for malformed URI encoding + TODO for lifecycle tests

Changes:
- Created packages/core/src/internal-api.ts (362 lines)
- Reduced packages/core/src/server.ts from 662 to 421 lines
- Added new test case for malformed URI handling
- Exported mountInternalApi, TimelineEntry, InternalApiDeps from index

Addresses: code-review-findings
